SoundMedScope/StorageMS/requirements.txt
Flask==3.0.2
paho-mqtt
python-dotenv
psycopg2-binary
boto3